requirements.txt
Django==4.2.7
djangorestframework==3.14.0
celery==5.3.4
redis==5.0.1
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
django-cors-headers==4.3.1
django-filter==23.3
Pillow==10.1.0
beautifulsoup4==4.12.2
requests==2.31.0
pandas==2.1.3
openpyxl==3.1.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0

